Skin issue fix, anim items
New function in groundchestitemtiles.lua for creating an animated item display.
searchFunction.lua GenerateItemList will now also get the item's bank, build and animation.
Implemented a function in searchFunction.lua for getting "true skin name" for when an atlas for a tex isn't found.
Made groundchestpickupqueuer and highlighting work with the new "GetTrueSkinName" function.
Items, which are selected to be picked up as "All" will now have a different colour(Yellow to be precise) than items  chosen to be picked up closest.