Changed closest item method
- groundchestpickupqueuer will now find the closest item by going through the item list and finding the smallest distance.
Note: table sorting proves very ineffective due to the fact that it requires more calculations than there are items in the list. This can cause lag spikes whenever the item is picked up and isn't generally a wanted method as the table will get sorted for as many items as there are.